add jwt auth to all pages and api endpoints that give event or ride information

+++ src/routes/api/carpool/vehicle/[id]/+server.js
@@ -1,7 +1,11 @@
 import { json } from '@sveltejs/kit'
 import Vehicle from '$lib/server/vehicle.js'
+import User from '$lib/server/user'
 
-export async function DELETE({ params }) {
+export async function DELETE({ params, cookies }) {
+  const jwt = cookies.get('jwt')
+  if (!jwt || !User.validate(jwt)) return new Response('Unauthorized', { status: 401 })
+
   try {
     await Vehicle.deleteVehicle(params.id)
     return json({ success: true })
